- Moves logic for restoring backups to its own module - Refactors `restore_backup()` method into `with_backup()` method that should also spawn the backup task
